fixed some bugs, have balls bouncing but they're not centered properly

+++ src/GameState.hs
@@ -21,7 +21,7 @@ import Tiles
 data GameData = GameData {
     masks :: ! Masks,
     player :: ! Entity,
-    physics :: ! (Physics Double),
+    physics :: ! (Physics Float),
     randomState :: ! StdGen,
     renderData :: ! (Vector.Vector RenderData),
     tiledMap :: ! TiledMap,
